+++ src/sevn/integrations/social_media/x_ops.py
@@ -33,6 +33,9 @@ __all__ = [
     "delete_bookmark",
     "delete_retweet",
     "delete_tweets",
+    "discover_followers",
+    "discover_mutual_graph",
+    "discover_topic_accounts",
     "fetch_article_markdown",
     "follow_user",
     "get_new_comments_on_tweet",
@@ -94,6 +97,9 @@ create_tweet_or_reply = _wrap("create_tweet_or_reply")
 create_quote_tweet = _wrap("create_quote_tweet")
 create_tweet_thread = _wrap("create_tweet_thread")
 delete_tweets = _wrap("delete_tweets")
+discover_followers = _wrap("discover_followers")
+discover_mutual_graph = _wrap("discover_mutual_graph")
+discover_topic_accounts = _wrap("discover_topic_accounts")
 post_tweet_auto_cookie = _wrap("post_tweet_auto_cookie")
 get_users_by_usernames = _wrap("get_users_by_usernames")
 follow_user = _wrap("follow_user")